job-search/requirements.txt
requests
fastapi
uvicorn
python-dotenv
google-generativeai>=0.7.0